Rename class "MoveImageScreen" to "MenuScreen".
