+++ core/src/ru/geekbrains/android/Star2DShooter.java
@@ -2,12 +2,12 @@ package ru.geekbrains.android;
 
 import com.badlogic.gdx.Game;
 
-import ru.geekbrains.android.screen.MoveImageScreen;
+import ru.geekbrains.android.screen.MenuScreen;
 
 
 public class Star2DShooter extends Game {
     @Override
     public void create() {
-        setScreen(new MoveImageScreen());
+        setScreen(new MenuScreen());
     }
 }
